requirements.txt
Click==7.0
dnspython==1.16.0
eventlet==0.25.0
Flask==1.0.3
Flask-SocketIO==4.1.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
psycopg2==2.8.2
python-engineio==3.8.1
python-socketio==4.1.0
PyYAML==5.1.1
redis==3.2.1
six==1.12.0
Werkzeug==0.15.4
